Refactoring Trainer helper classes (WIP)
